feat: 新トップ(旧/dev/top5)を本番トップ / に昇格
- src/app/page.tsx: 未ログイン時に top-next organisms 構成の新トップを表示。
  ログイン済みは従来どおり /mypage リダイレクト。本番メタ（title/desc/OG/
  canonical/indexable）とWebSite JSON-LDを維持。到達は未ログインのみのため
  Hero入会CTAは常時表示(isMember=false)
- Layout.tsx: `/`(home) を half-gradient 分岐に追加し新トップの見た目に合わせる
- /dev/top5 と孤立した DevNav.tsx を削除（比較用 /dev/top6 は残置）
- dev/page.tsx: 索引の top5 エントリを top6 に置換

検証: build成功 / tsc通過 / 全内部リンクが実在ルートに解決 /
見出し h1(1)→h2→h3 階層OK / メタ・OGP(1200x630)・canonical・indexable 確認
